refactor and review
